Switched to slash commands

+++ utilities.go
@@ -1,7 +1,10 @@
 package main
 
 import (
+	"crypto/sha1"
 	"database/sql"
+	"encoding/base32"
+	"errors"
 	"github.com/bwmarrin/discordgo"
 	"github.com/bwmarrin/lit"
 	"math/rand"
@@ -10,27 +13,21 @@ import (
 	"runtime"
 	"strings"
 	"sync"
+	"time"
 )
 
 const (
 	tblCustomCommands = "CREATE TABLE IF NOT EXISTS \"customCommands\" (\"server\" VARCHAR(18) NOT NULL,\"command\" VARCHAR(50) NOT NULL,\"text\" VARCHAR(2000) NOT NULL);"
 )
 
-// deleteMessage delete a message
-func deleteMessage(s *discordgo.Session, m *discordgo.Message) {
-	lit.Info("%s: %s", m.Author.Username, m.Content)
-	err := s.ChannelMessageDelete(m.ChannelID, m.ID)
-	if err != nil {
-		lit.Error("Can't delete message, %s", err)
-	}
-}
-
 // findUserVoiceState finds the voicestate of a user
-func findUserVoiceState(session *discordgo.Session, userID string) *discordgo.VoiceState {
-	for _, guild := range session.State.Guilds {
-		for _, vs := range guild.VoiceStates {
-			if vs.UserID == userID {
-				return vs
+func findUserVoiceState(s *discordgo.Session, userID string, guidID string) *discordgo.VoiceState {
+	for _, g := range s.State.Guilds {
+		if g.ID == guidID {
+			for _, vs := range g.VoiceStates {
+				if vs.UserID == userID {
+					return vs
+				}
 			}
 		}
 	}
@@ -65,12 +62,12 @@ func execQuery(query string, db *sql.DB) {
 }
 
 // Adds a custom command to db and to the command map
-func addCommand(command string, text string, guild string) {
+func addCommand(command string, text string, guild string) error {
 	initializeServer(guild)
 
 	// If the text is already in the map, we ignore it
 	if server[guild].customCommands[command] == text {
-		return
+		return errors.New("command already exists")
 	}
 
 	// Else, we add it to the map
@@ -82,24 +79,35 @@ func addCommand(command string, text string, guild string) {
 	_, err := stm.Exec(guild, command, text)
 	if err != nil {
 		lit.Error("Error inserting into the database, %s", err)
+		return errors.New("error inserting into the database: " + err.Error())
 	}
 
 	_ = stm.Close()
+
+	return nil
 }
 
 // Removes a custom command from the db and from the command map
-func removeCustom(command string, guild string) {
+func removeCustom(command string, guild string) error {
+
+	if server[guild].customCommands[command] == "" {
+		return errors.New("command doesn't exist")
+	}
+
 	// Remove from DB
 	stm, _ := db.Prepare("DELETE FROM customCommands WHERE server=? AND command=?")
 	_, err := stm.Exec(guild, command)
 	if err != nil {
 		lit.Error("Error removing from the database, %s", err)
+		return errors.New("error removing from the database: " + err.Error())
 	}
 
 	_ = stm.Close()
 
 	// Remove from the map
 	delete(server[guild].customCommands, command)
+
+	return nil
 }
 
 // Loads custom command from the database
@@ -199,4 +207,48 @@ func initializeServer(guildID string) {
 			customCommands: make(map[string]string),
 		}
 	}
+}
+
+// Sends embed as response to an interaction
+func sendEmbedInteraction(s *discordgo.Session, embed *discordgo.MessageEmbed, i *discordgo.Interaction) {
+	sliceEmbed := []*discordgo.MessageEmbed{embed}
+	err := s.InteractionRespond(i, &discordgo.InteractionResponse{Type: discordgo.InteractionResponseChannelMessageWithSource, Data: &discordgo.InteractionApplicationCommandResponseData{Embeds: sliceEmbed}})
+	if err != nil {
+		lit.Error("InteractionRespond failed: %s", err)
+		return
+	}
+}
+
+// Sends and delete after three second an embed in a given channel
+func sendAndDeleteEmbedInteraction(s *discordgo.Session, embed *discordgo.MessageEmbed, i *discordgo.Interaction, wait time.Duration) {
+	sendEmbedInteraction(s, embed, i)
+
+	time.Sleep(wait)
+
+	err := s.InteractionResponseDelete(s.State.User.ID, i)
+	if err != nil {
+		lit.Error("InteractionResponseDelete failed: %s", err)
+		return
+	}
+}
+
+// Modify an already sent interaction
+func modfyInteraction(s *discordgo.Session, embed *discordgo.MessageEmbed, i *discordgo.Interaction) {
+	sliceEmbed := []*discordgo.MessageEmbed{embed}
+	err := s.InteractionResponseEdit(s.State.User.ID, i, &discordgo.WebhookEdit{Embeds: sliceEmbed})
+	if err != nil {
+		lit.Error("InteractionResponseEdit failed: %s", err)
+		return
+	}
+}
+
+// genAudio generates a dca file from a string
+func genAudio(text string) string {
+	h := sha1.New()
+	h.Write([]byte(text))
+	uuid := strings.ToUpper(base32.HexEncoding.EncodeToString(h.Sum(nil)))
+
+	gen(text, uuid)
+
+	return uuid + ".dca"
 }
